Learned IMV version of non partition join

+++ src/benchmark_runners/non_partition_join_runner.cpp
@@ -450,15 +450,15 @@ void npj_build_rel_r_partition_learned(ETHNonPartitionJoinBuild<KeyType, Payload
         if (prefetch_index < rel_r_partition->num_tuples) {
             idx_prefetch = static_cast<uint64_t>(std::max(
                                 0.,
-                            std::min(num_models - 1., root_slope * (double)(rel_r_partition->tuples[prefetch_index].key) + root_intrcpt)));
+                            std::min(num_models - 1., root_slope * rel_r_partition->tuples[prefetch_index].key + root_intrcpt)));
 
             // Predict the CDF
             pred_cdf =
-                (*slopes)[idx_prefetch] * (double)(rel_r_partition->tuples[prefetch_index].key) + (*intercepts)[idx_prefetch];
+                (*slopes)[idx_prefetch] * rel_r_partition->tuples[prefetch_index].key + (*intercepts)[idx_prefetch];
 
             // Scale the CDF to the number of buckets
             idx_prefetch = static_cast<uint64_t>(
-                std::max(0., std::min(FANOUT - 1., pred_cdf * (double)FANOUT)));    
+                std::max(0., std::min(FANOUT - 1., pred_cdf * FANOUT)));    
             
             prefetch_index++;
 			__builtin_prefetch(ht->buckets + idx_prefetch, 1, 1);
@@ -467,15 +467,15 @@ void npj_build_rel_r_partition_learned(ETHNonPartitionJoinBuild<KeyType, Payload
 
         idx = static_cast<uint64_t>(std::max(
                                 0.,
-                               std::min(num_models - 1., root_slope * (double)rel_r_partition->tuples[i].key + root_intrcpt)));
+                               std::min(num_models - 1., root_slope * rel_r_partition->tuples[i].key + root_intrcpt)));
 
         // Predict the CDF
         pred_cdf =
-            (*slopes)[idx] * (double)rel_r_partition->tuples[i].key + (*intercepts)[idx];
+            (*slopes)[idx] * rel_r_partition->tuples[i].key + (*intercepts)[idx];
 
         // Scale the CDF to the number of buckets
         idx = static_cast<uint64_t>(
-            std::max(0., std::min(FANOUT - 1., pred_cdf * (double)FANOUT)));
+            std::max(0., std::min(FANOUT - 1., pred_cdf * FANOUT)));
 
         /*if(rel_r_partition->tuples[i].key < 10)
         {
@@ -1247,15 +1247,15 @@ uint64_t npj_probe_rel_s_partition_learned(Relation<KeyType, PayloadType> * rel_
         {
             idx_prefetch = static_cast<uint64_t>(std::max(
                                 0.,
-                            std::min(num_models - 1., root_slope * (double)rel_s_partition->tuples[prefetch_index].key + root_intrcpt)));
+                            std::min(num_models - 1., root_slope * rel_s_partition->tuples[prefetch_index].key + root_intrcpt)));
 
             // Predict the CDF
             pred_cdf =
-                (*slopes)[idx_prefetch] * (double)rel_s_partition->tuples[prefetch_index].key + (*intercepts)[idx_prefetch];
+                (*slopes)[idx_prefetch] * rel_s_partition->tuples[prefetch_index].key + (*intercepts)[idx_prefetch];
 
             // Scale the CDF to the number of buckets
             idx_prefetch = static_cast<uint64_t>(
-                std::max(0., std::min(FANOUT - 1., pred_cdf * (double)FANOUT)));    
+                std::max(0., std::min(FANOUT - 1., pred_cdf * FANOUT)));    
 
             prefetch_index++;
 
@@ -1265,24 +1265,24 @@ uint64_t npj_probe_rel_s_partition_learned(Relation<KeyType, PayloadType> * rel_
         
         idx = static_cast<uint64_t>(std::max(
                                 0.,
-                            std::min(num_models - 1., root_slope * (double)rel_s_partition->tuples[i].key + root_intrcpt)));
+                            std::min(num_models - 1., root_slope * rel_s_partition->tuples[i].key + root_intrcpt)));
 
         // Predict the CDF
         pred_cdf =
-            (*slopes)[idx] * (double)rel_s_partition->tuples[i].key + (*intercepts)[idx];
+            (*slopes)[idx] * rel_s_partition->tuples[i].key + (*intercepts)[idx];
 
         // Scale the CDF to the number of buckets
         idx = static_cast<uint64_t>(
-            std::max(0., std::min(FANOUT - 1., pred_cdf * (double)FANOUT)));
+            std::max(0., std::min(FANOUT - 1., pred_cdf * FANOUT)));
 
         if(i < 10)
         {
             printf("key %ld root_slope %f root_intrcpt %f root_slope * rel_s_partition->tuples[i].key + root_intrcpt %f idx_first %ld (*slopes)[idx] %f (*intercepts)[idx] %f pred_cdf %f pred_cdf * FANOUT %f idx %ld \n", 
-                    rel_s_partition->tuples[i].key, root_slope, root_intrcpt, root_slope * (double)rel_s_partition->tuples[i].key + root_intrcpt, 
+                    rel_s_partition->tuples[i].key, root_slope, root_intrcpt, root_slope * rel_s_partition->tuples[i].key + root_intrcpt, 
                     static_cast<uint64_t>(std::max(
                                 0.,
-                            std::min(num_models - 1., root_slope * (double)rel_s_partition->tuples[i].key + root_intrcpt))),
-                    (*slopes)[idx], (*intercepts)[idx], pred_cdf, pred_cdf * (double)FANOUT, idx);
+                            std::min(num_models - 1., root_slope * rel_s_partition->tuples[i].key + root_intrcpt))),
+                    (*slopes)[idx], (*intercepts)[idx], pred_cdf, pred_cdf * FANOUT, idx);
         }    
 
         Bucket<KeyType, PayloadType> * b = ht->buckets+idx;
